Finished grid class
A little tired, I finished the grid class and forgot to push to repository.  I started a bit of work on other sections which is detailed in log.  Normally I wouldn't half-step logic like this, but the grid was enough work that I don't want to lose.  processors.py is broken in this build.